kill count, coin collect and enemy health bug
The kill count was set to 0 in every scene, it was difficult to collect coins, collider was re-adjusted and we died when dead enemies hit us.

+++ Assets/Scripts/Shooting/BaseShooter.cs
@@ -3,8 +3,8 @@ using UnityEngine;
 public class BaseShooter : MonoBehaviour
 {
     [SerializeField] protected float shootSpeed;
-    [SerializeField] protected int shootingStartDelay;
-    [SerializeField] protected int shootingInterval;
+    [SerializeField] protected float shootingStartDelay;
+    [SerializeField] protected float shootingInterval;
 
     [SerializeField] protected Transform spawnPointOfProjectile;
     [SerializeField] protected GameObject projectile;
